refactor: add package controller and model
