+++ src/main/java/gift/product/controller/ProductController.java
@@ -1,10 +1,10 @@
-package gift.product;
+package gift.product.controller;
 
+import gift.product.model.ProductVo;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
